declare customer matcher types
